+++ aqua/core/intake_drivers/__init__.py
@@ -1,8 +1,7 @@
 import intake  # Import this first to avoid circular imports during discovery.
 
 # from intake.container import register_container
-from .fdb import IntakeFDBSource
-from .openers import open_gsv, open_polytope
+from .fdb import IntakeFDBSource, open_gsv, open_polytope
 
 __all__ = ["IntakeFDBSource", "open_gsv", "open_polytope"]
