+++ aqua/core/intake_drivers/fdb/fdb.py
@@ -1,6 +1,6 @@
-from aqua.core.fdb.intake2.base import IntakeFDBSourceAdapter
-from aqua.core.fdb.intake2.datatypes import GSV, Polytope
-from aqua.core.fdb.intake2.readers import GSVDatasetReader, PolytopeDatasetReader
+from .base import IntakeFDBSourceAdapter
+from .datatypes import GSV, Polytope
+from .readers import GSVDatasetReader, PolytopeDatasetReader
 
 
 class IntakeFDBSource(IntakeFDBSourceAdapter):
